Summary: Tracking blood sugar now works as expected. (In the same way as the symptom tracking works.)
Adds BloodSugar screens for tracking blood sugar, and recording entered blood surgar values on the main screen.

Fixes an issue with the PrefSingleton LoadPreferenceList function only
working with lists of the Symptom type (now works with BloodSugar)
The PrefSingleton should be completely working properly now.

Also adds more strings to the @string resource, so translations can be
made more easily in the future.